back to shared ctxs with option to switch for testing

+++ test/wallet/action/internalizeAction.test.ts
@@ -7,16 +7,28 @@ describe('internalizeAction tests', () => {
 
   const env = _tu.getEnv('test')
 
+  const gctxs: TestWalletNoSetup[] = []
+  const useSharedCtxs = true
+
   beforeAll(async () => {
+    if (!env.noMySQL) gctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeActionTests'))
+    gctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeActionTests'))
   })
 
   afterAll(async () => {
+    for (const ctx of gctxs) {
+      await ctx.storage.destroy()
+    }
   })
 
   test('1 internalize custom output in receiving wallet with checks', async () => {
     const ctxs: TestWalletNoSetup[] = []
-    if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction1Tests'))
-    ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction1Tests'))
+    if (useSharedCtxs)
+      ctxs.push(...gctxs)
+    else {
+      if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction1Tests'))
+      ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction1Tests'))
+    }
     for (const { wallet } of ctxs) {
       const root = '02135476'
       const kp = _tu.getKeyPair(root.repeat(8))
@@ -81,15 +93,21 @@ describe('internalizeAction tests', () => {
         await fred.activeStorage.destroy()
       }
     }
-    for (const ctx of ctxs) {
-      await ctx.storage.destroy()
+    if (!useSharedCtxs) {
+      for (const ctx of ctxs) {
+        await ctx.storage.destroy()
+      }
     }
   })
 
   test('2 internalize 2 custom outputs in receiving wallet with checks', async () => {
     const ctxs: TestWalletNoSetup[] = []
-    if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction2Tests'))
-    ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction2Tests'))
+    if (useSharedCtxs)
+      ctxs.push(...gctxs)
+    else {
+      if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction2Tests'))
+      ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction2Tests'))
+    }
     for (const { wallet } of ctxs) {
       const root = '02135476'
       const kp = _tu.getKeyPair(root.repeat(8))
@@ -181,15 +199,21 @@ describe('internalizeAction tests', () => {
         await fred.activeStorage.destroy()
       }
     }
-    for (const ctx of ctxs) {
-      await ctx.storage.destroy()
+    if (!useSharedCtxs) {
+      for (const ctx of ctxs) {
+        await ctx.storage.destroy()
+      }
     }
   })
 
   test('3 internalize wallet payment in receiving wallet with checks', async () => {
     const ctxs: TestWalletNoSetup[] = []
-    if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction3Tests'))
-    ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction3Tests'))
+    if (useSharedCtxs)
+      ctxs.push(...gctxs)
+    else {
+      if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction3Tests'))
+      ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction3Tests'))
+    }
     for (const { wallet, identityKey: senderIdentityKey } of ctxs) {
       const fred = await _tu.createSQLiteTestWallet({ chain: 'test', databaseName: 'internalizeAction3fred', rootKeyHex: '2'.repeat(64), dropAll: true })
       const outputSatoshis = 5
@@ -254,15 +278,21 @@ describe('internalizeAction tests', () => {
         await fred.activeStorage.destroy()
       }
     }
-    for (const ctx of ctxs) {
-      await ctx.storage.destroy()
+    if (!useSharedCtxs) {
+      for (const ctx of ctxs) {
+        await ctx.storage.destroy()
+      }
     }
   })
 
   test('4 internalize 2 wallet payments in receiving wallet with checks', async () => {
     const ctxs: TestWalletNoSetup[] = []
-    if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction4Tests'))
-    ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction4Tests'))
+    if (useSharedCtxs)
+      ctxs.push(...gctxs)
+    else {
+      if (!env.noMySQL) ctxs.push(await _tu.createLegacyWalletMySQLCopy('actionInternalizeAction4Tests'))
+      ctxs.push(await _tu.createLegacyWalletSQLiteCopy('actionInternalizeAction4Tests'))
+    }
     for (const { wallet, identityKey: senderIdentityKey } of ctxs) {
       const fred = await _tu.createSQLiteTestWallet({ chain: 'test', databaseName: 'internalizeAction4fred', rootKeyHex: '2'.repeat(64), dropAll: true })
 
@@ -351,8 +381,10 @@ describe('internalizeAction tests', () => {
         await fred.activeStorage.destroy()
       }
     }
-    for (const ctx of ctxs) {
-      await ctx.storage.destroy()
+    if (!useSharedCtxs) {
+      for (const ctx of ctxs) {
+        await ctx.storage.destroy()
+      }
     }
   })
